# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HackR API",
        "description": "Here is my Swagger documentation for my HackR API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/passwords/check": {
            "post": {
                "tags": [
                    "Fonctionnalités"
                ],
                "summary": "Check password strength",
                "description": "Check if a password is common from a predefined list.",
                "operationId": "c32c8b9b701c1a228de765c432d4e0a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "The password to check if it is common.",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The password is common"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Password list file not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Password list file not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The password field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. User lacks sufficient permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "You do not have permission to perform this action."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/passwords/generate": {
            "get": {
                "tags": [
                    "Fonctionnalités"
                ],
                "summary": "Generate a Secure Password",
                "description": "Generates a secure password containing at least one uppercase letter, one lowercase letter, one number, and one special character. The default password length is 12 characters, but you can specify a different length.",
                "operationId": "generateSecurePassword",
                "parameters": [
                    {
                        "name": "length",
                        "in": "query",
                        "description": "The length of the generated password (default is 12 if not provided)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 12,
                            "example": 16
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully generated secure password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "password": {
                                            "description": "The generated secure password",
                                            "type": "string"
                                        },
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User does not have the right to generate a password",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/phishing/data": {
            "post": {
                "tags": [
                    "Fonctionnalités"
                ],
                "summary": "Phishing",
                "description": "Phishing attack to capture user credentials.",
                "operationId": "77c31203b6e9ec986904a363c139064c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "description": "The URL of the webpage to inject the script.",
                                        "type": "string",
                                        "example": "https://fr-fr.facebook.com/login/web/"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HTML page with the phishing script injected.",
                        "content": {
                            "text/html": {
                                "example": "<html><head></head><body>...</body></html>"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request. Validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The url field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}